yaml: set PREFERRED_PROVIDER for virtual/libgles3
There are some packages (like kmscube) which require
libgles3. Imagination DDK actually provides support for this standard,
but for some reason it was not advertised in recipes provided by
Renesas. There is a patch to meta-xt-rcar that sets PROVIDE variable,
and there we need to chose correct PREFERRED_PROVIDER in the same way
as we did for libgles2.

Signed-off-by: Volodymyr Babchuk <volodymyr_babchuk@epam.com>